storage: switch Store.{replicas,replicaQueues} from sync.Map to IntMap
Motivated by seeing sync.Map.Load() high on profiles when there are
large numbers of replicas on a node. See #17609.

name               old time/op  new time/op  delta
StoreGetReplica-8  9.12ns ±10%  3.22ns ± 5%  -64.65%  (p=0.000 n=10+8)

+++ pkg/storage/replica.go
@@ -4341,9 +4341,9 @@ func (r *Replica) acquireSplitLock(
 			rightRng.mu.destroyed = errors.Errorf("%s: failed to initialize", rightRng)
 			rightRng.mu.Unlock()
 			r.store.mu.Lock()
-			r.store.mu.replicas.Delete(rightRng.RangeID)
+			r.store.mu.replicas.Delete(int64(rightRng.RangeID))
 			delete(r.store.mu.uninitReplicas, rightRng.RangeID)
-			r.store.replicaQueues.Delete(rightRng.RangeID)
+			r.store.replicaQueues.Delete(int64(rightRng.RangeID))
 			r.store.mu.Unlock()
 		}
 		rightRng.raftMu.Unlock()
